+++ package.json
@@ -15,7 +15,8 @@
     "react-dom": "^16.4.0",
     "react-redux": "^5.0.7",
     "react-select": "^1.2.1",
-    "redux": "^4.0.0"
+    "redux": "^4.0.0",
+    "reselect": "^3.0.1"
   },
   "devDependencies": {
     "babel-core": "^6.13.2",
